making InMemoryStorage an abstract DocumentBinaryStorage .. and implementing it in-memory
